Play-Test no depending on Akka HTTP and Netty servers
