+++ dev-mode/sbt-plugin/src/sbt-test/play-sbt-plugin/http-backend-system-property/build.sbt
@@ -4,7 +4,7 @@
 
 lazy val root = (project in file(".")).enablePlugins(PlayScala).enablePlugins(MediatorWorkaroundPlugin)
 
-name := "system-property"
+name := "http-backend-system-property"
 
 scalaVersion := sys.props("scala.version")
 updateOptions := updateOptions.value.withLatestSnapshots(false)
@@ -13,10 +13,7 @@ evictionWarningOptions in update ~= (_.withWarnTransitiveEvictions(false).withWa
 // because the "test" directory clashes with the scripted test file
 scalaSource in Test := (baseDirectory.value / "tests")
 
-val playAkkaHttpServer = "com.typesafe.play" %% "play-akka-http-server" % sys.props("project.version")
-val playNettyServer    = "com.typesafe.play" %% "play-netty-server"     % sys.props("project.version")
-
-libraryDependencies ++= Seq(playAkkaHttpServer, playNettyServer, guice, ws, specs2 % Test)
+libraryDependencies ++= Seq(akkaHttpServer, nettyServer, guice, ws, specs2 % Test)
 
 fork in Test := true
