+++ dev-mode/sbt-plugin/src/sbt-test/play-sbt-plugin/http-backend-system-property/test
@@ -1,8 +1,3 @@
-# Start dev mode with the default server - AkkaHttpServer
-> run
-> verifyResourceContains / 200 unknown
-> playStop
-
 # Start dev mode with an overridden server - AkkaHttpServer
 > run -Dplay.server.provider=play.core.server.AkkaHttpServerProvider
 > verifyResourceContains / 200 unknown
